back/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pymysql==1.1.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
reportlab==4.2.5
